requirements.txt
fastapi
uvicorn
sqlalchemy
pymupdf
google-generativeai
python-dotenv
python-multipart
aiofiles